Перерабатывает TasksPage.
