+++ src/pages/tasks/TasksPage.style.ts
@@ -2,12 +2,10 @@ import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
 
 export default makeStyles((theme: Theme) =>
   createStyles({
-    drawer: {
-      width: '240',
-    },
-    grid: {
-      marginTop: theme.spacing(2),
-      marginBottom: theme.spacing(2),
+    fab: {
+      position: 'fixed',
+      bottom: theme.spacing(4),
+      right: theme.spacing(4),
     },
   })
 );
